requirements.txt
google-generativeai
python-dotenv
langchain
langchain-community
langchain-google-genai
fastapi
uvicorn
numpy
pandas
faiss-cpu
